# packages/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: AVANA API
  version: 1.0.0
  description: AVANA API contract (Sprint 1 foundations)
servers:
  - url: https://api.example.invalid
paths:
...
  /v1/organizations/{organizationId}/courses/{courseId}/generated/review-queue:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
    get:
      operationId: getGeneratedReviewQueue
      summary: List AI-generated content pending review for a course (draft/edited)
      responses:
        "200":
          description: Review queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewQueueResponse"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Course not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/generated/{contentId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/GeneratedContentId"
    get:
      operationId: getGeneratedContentForReview
      summary: Get a single generated content with citations, source chunks, metadata, and generation info for review
      responses:
        "200":
          description: Generated content review view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContentReviewResource"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Generated content not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: editGeneratedContent
      summary: Edit a generated content before acceptance (editor modification)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditGeneratedContentRequest"
      responses:
        "200":
          description: Edited generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContentReviewResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content editing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Generated content not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Content is not in an editable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/generated/{contentId}/accept:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/GeneratedContentId"
    post:
      operationId: acceptGeneratedContent
      summary: Accept a generated content (draft → accepted). Materializes a lesson into the Learning Core.
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptContentResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content acceptance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Generated content not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Content is not in an acceptable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/generated/{contentId}/reject:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/GeneratedContentId"
    post:
      operationId: rejectGeneratedContent
      summary: Reject a generated content (draft/edited → rejected). Requires a rejection reason.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectContentRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RejectContentResponse"
        "400":
          description: Rejection reason is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content rejection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Generated content not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Content is not in a rejectable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/organizations/{organizationId}/courses/{courseId}/generated/{contentId}/regenerate:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/GeneratedContentId"
    post:
      operationId: regenerateGeneratedContent
      summary: Request async regeneration of a generated content
      responses:
        "202":
          description: Regeneration accepted (async job)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegenerateContentResponse"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role does not permit content regeneration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Generated content not found or not visible to the actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Content is already being regenerated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes: {}
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    GeneratedContentId:
      name: contentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: ["error", "request_id"]
      properties:
        request_id:
          type: string
        error:
          type: object
          additionalProperties: false
          required: ["code", "message"]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - unprocessable
                - internal_error
            message:
              type: string
            details:
              type: object
              additionalProperties:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
                  - type: integer
                  - type: null
...
    GeneratedContentType:
      type: string
      description: Extensible union of AI-generated content types. Only "lesson" is enabled in PR6-4; flashcard/quiz/recommendation are added in later PRs.
      enum:
        - lesson
        - flashcard
        - quiz
        - recommendation
    GeneratedContentStatus:
      type: string
      description: AI artifact lifecycle (separate from the document processing lifecycle).
      enum:
        - draft
        - accepted
        - rejected
        - edited
        - regenerating
    CitationResource:
      type: object
      additionalProperties: false
      required: ["document_chunk_id"]
      properties:
        document_chunk_id:
          type: string
          format: uuid
    TokenUsage:
      type: object
      additionalProperties: false
      required: ["input_tokens", "output_tokens"]
      properties:
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
    GeneratedContentResource:
      type: object
      additionalProperties: false
      required:
        - id
        - document_id
        - course_id
        - type
        - status
        - payload
        - prompt_version
        - model
        - token_usage
        - citations
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/GeneratedContentType"
        status:
          $ref: "#/components/schemas/GeneratedContentStatus"
        payload:
          type: object
          additionalProperties: true
          description: Typed content payload (discriminated by type).
        prompt_version:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        token_usage:
          $ref: "#/components/schemas/TokenUsage"
        citations:
          type: array
          items:
            $ref: "#/components/schemas/CitationResource"
        created_at:
          type: string
          description: UTC ISO datetime
        updated_at:
          type: string
          description: UTC ISO datetime
...
    ReviewQueueResource:
      type: object
      additionalProperties: false
      required: ["id", "document_id", "course_id", "type", "status", "title", "updated_at"]
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/GeneratedContentType"
        status:
          $ref: "#/components/schemas/GeneratedContentStatus"
        title:
          type: string
        updated_at:
          type: string
          description: UTC ISO datetime
    ReviewQueueResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "pending"]
      properties:
        request_id:
          type: string
        pending:
          type: array
          items:
            $ref: "#/components/schemas/ReviewQueueResource"
    SourceChunkResource:
      type: object
      additionalProperties: false
      required: ["id", "sequence", "heading", "content", "start_page", "end_page"]
      properties:
        id:
          type: string
          format: uuid
        sequence:
          type: integer
        heading:
          type: string
          nullable: true
        content:
          type: string
        start_page:
          type: integer
        end_page:
          type: integer
    GenerationInfoResource:
      type: object
      additionalProperties: false
      required: ["model", "prompt_version", "token_usage"]
      properties:
        model:
          type: string
          nullable: true
        prompt_version:
          type: string
          nullable: true
        token_usage:
          $ref: "#/components/schemas/TokenUsage"
    GeneratedContentReviewResource:
      type: object
      additionalProperties: false
      required: ["request_id", "content", "source_chunks", "generation"]
      properties:
        request_id:
          type: string
        content:
          $ref: "#/components/schemas/GeneratedContentResource"
        source_chunks:
          type: array
          items:
            $ref: "#/components/schemas/SourceChunkResource"
        generation:
          $ref: "#/components/schemas/GenerationInfoResource"
    GeneratedContentReviewResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "content", "source_chunks", "generation"]
      properties:
        request_id:
          type: string
        content:
          $ref: "#/components/schemas/GeneratedContentResource"
        source_chunks:
          type: array
          items:
            $ref: "#/components/schemas/SourceChunkResource"
        generation:
          $ref: "#/components/schemas/GenerationInfoResource"
    EditGeneratedContentRequest:
      type: object
      additionalProperties: false
      required: ["payload"]
      properties:
        payload:
          type: object
          additionalProperties: true
          description: The edited typed content payload.
    AcceptContentResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "content_id", "status", "materialized_lesson_id"]
      properties:
        request_id:
          type: string
        content_id:
          type: string
          format: uuid
        status:
          type: string
          const: accepted
        materialized_lesson_id:
          type: string
          format: uuid
          nullable: true
    RejectContentRequest:
      type: object
      additionalProperties: false
      required: ["reason"]
      properties:
        reason:
          type: string
          minLength: 1
    RejectContentResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "content_id", "status"]
      properties:
        request_id:
          type: string
        content_id:
          type: string
          format: uuid
        status:
          type: string
          const: rejected
    RegenerateContentResponse:
      type: object
      additionalProperties: false
      required: ["request_id", "content_id", "job_id", "status"]
      properties:
        request_id:
          type: string
        content_id:
          type: string
          format: uuid
        job_id:
          type: string
        status:
          type: string
          const: regenerating
...